pyproject.toml
[project]
name = "payu-main-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
